Fix definition of Result.baseline
previously the build to compare was defined in terms of the baseline,
but that is not correct becasue the baseline was not being obtained when
it had no results. Whether the baseline build contains results or not
should not interefere with which build if the baseline of a given build.