benchmarks.set_testjob_result: don't serialize TestJob objects
Every time we upgrade and there are changes to the TestJob class, the
objects that were serialized before the upgrade are outdated wrt the
code, and we get errors becasue the code tries to access attributes that
don't exist.

Instead of serializing the entire object, let's serialized just its id,
and pull a fresh object out of the DB exactly when we need it.