Recreate benchmarks result data
I wondered how to associate the TestJobs with the existing ResultData
objects, but couldn't figure out how. Even if we matched them based on
the _64 or _32 prefixes that the ResultData objects had, that would
still not capture the difference between different core cluster (e.g.
A53 x A57).

So with this we are losing a lot of data, but I'm afraid that is data
that can't be put to good use anyway.