API: avoid background tasks looking for uncommitted data
Since a POST to /api/results is atomic, if the background tasks are
processed before the submission finished, the necessary data will not be
committed to the database yet. Scheduling those tasks to 1 minute after
all processing has finished should give them enough time to get a
consistent view of the database.