benchmarks: removed assumption that baseline comes before patch
Baseline builds might be reported after the patched build is already in
the database. In this case email reporting will show "no baseline".
However after the baseline build is introduced to database, the detailed
view should use it to compare test results. So assumption that baseline
comes before patch is wrong.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>