API: fixed the simultaneus requests problem
When there were two requests coming at the same time trying to create a
Result object with the same data, it was possible to create two
identical objects. Since this didn't break database integrity it was
allowed. However the UI would present the results as two separate
'builds'. To prevent that 'build_id' and 'name' has to be unique
together. Now trying to create more than one Result object will raise
IntegrityError. This exception is handled in the api/views.py so the
outside behaviour for the user doesn't change.

Signed-off-by: Milosz Wasilewski <milosz.wasilewski@linaro.org>